Fix upgrade leaving services stopped. Minor layout change in ncpa.cfg.

+++ build/resources/ncpa-partial.spec
@@ -26,11 +26,11 @@ bundled version of Python.
 rm -rf %{buildroot} 
 mkdir -p %{buildroot}/usr/local/ncpa
 mkdir -p %{buildroot}/usr/local/ncpa/var/run
-mkdir -p %{buildroot}/etc/init.d/
-touch %{buildroot}/usr/local/ncpa/ncpa.crt
-touch %{buildroot}/usr/local/ncpa/ncpa.key
+touch %{buildroot}/usr/local/ncpa/var/ncpa.crt
+touch %{buildroot}/usr/local/ncpa/var/ncpa.key
+touch %{buildroot}/usr/local/ncpa/var/ncpa.db
 cp -rf $RPM_BUILD_DIR/ncpa-%{version}/* %{buildroot}/usr/local/ncpa/
-chown nagios:nagios %{buildroot}/usr/local/ncpa -R
+chown -R nagios:nagios %{buildroot}/usr/local/ncpa
 install -m 755 $RPM_BUILD_DIR/ncpa-%{version}/build_resources/listener_init %{buildroot}/etc/init.d/ncpa_listener
 install -m 755 $RPM_BUILD_DIR/ncpa-%{version}/build_resources/passive_init %{buildroot}/etc/init.d/ncpa_passive
 
@@ -93,21 +93,38 @@ else
 fi
 
 %preun
-if [ `command -v systemctl` ]; then
-    systemctl stop ncpa_listener
-    systemctl stop ncpa_passive
-else
-    service ncpa_listener stop
-    service ncpa_passive stop
+# Only stop on actual uninstall not upgrades
+if [ "$1" != "1" ]; then
+    if [ `command -v systemctl` ]; then
+        systemctl stop ncpa_listener
+        systemctl stop ncpa_passive
+    else
+        service ncpa_listener stop
+        service ncpa_passive stop
+    fi
 fi
 
 %files
-%defattr(0755,root,root,-)
-/etc/init.d/ncpa_listener
-/etc/init.d/ncpa_passive
+%defattr(0755,nagios,nagios,-)
+%dir /usr/local/ncpa
+%dir /usr/local/ncpa/etc
+%dir /usr/local/ncpa/etc/ncpa.cfg.d
+/usr/local/ncpa/ncpa_listener
+/usr/local/ncpa/ncpa_passive
 
-%defattr(0775,nagios,nagios,-)
-/usr/local/ncpa
+%defattr(0644,nagios,nagios,-)
+/usr/local/ncpa/*.so
+/usr/local/ncpa/*.py
+/usr/local/ncpa/*.zip
 
+%defattr(0755,nagios,nagios,-)
+/usr/local/ncpa/build_resources
+/usr/local/ncpa/listener
+/usr/local/ncpa/plugins
+/usr/local/ncpa/var
+
+%defattr(0644,nagios,nagios,-)
 %config(noreplace) /usr/local/ncpa/etc/ncpa.cfg
 %config(noreplace) /usr/local/ncpa/etc/ncpa.cfg.d/example.cfg
+/usr/local/ncpa/etc/ncpa.cfg.sample
+/usr/local/ncpa/etc/ncpa.cfg.d/README.txt
